Refactor fonts: support proper referential scaling

+++ core/src/main/kotlin/io/github/chrislo27/paintbox/font/FontCache.kt
@@ -24,7 +24,7 @@ open class FontCache(val game: PaintboxGame) : Disposable {
         }
     }
 
-    fun resizeAll(width: Float, height: Float) {
+    fun resizeAll(width: Int, height: Int) {
         fonts.values.forEach {
             it.resize(width, height)
         }
